+++ goldsberry/__init__.py
@@ -9,3 +9,4 @@ import goldsberry.league
 import goldsberry.player
 import goldsberry.team
 from goldsberry.player._Player import PlayerList
+from goldsberry.game._Game import GameIDs